Refactor main.py and optimizer.py
- Added a debug check for trading symbols in main.py to ensure configuration integrity.
- Imported SYMBOLS from config in main.py for better configuration management.
- Updated import statement in optimizer.py to reflect the correct module path for Backtester.
