+++ configs/bot_config.yaml
@@ -1,7 +1,7 @@
 # Professional Trading Bot Configuration
 
 trading:
-  symbols: ['EURUSD', 'GBPUSD', 'USDJPY']
+  symbols: ['EURUSD','GBPUSD','XAUUSD']  # Fixed symbols for trading
   risk_per_trade: 1.0              # Risk 1% per trade
   max_position_size: 10.0          # Maximum 10% of account per position
   max_daily_risk: 5.0              # Maximum 5% daily risk
