+++ logs/performance_data.json
@@ -26,5 +26,5 @@
   },
   "strategy_performance": {},
   "symbol_performance": {},
-  "last_saved": "2025-07-29T04:26:26.680674"
+  "last_saved": "2025-07-29T10:08:04.219333"
 }
