+++ utils/optimizer.py
@@ -20,7 +20,7 @@ import itertools
 import pandas as pd
 from types import SimpleNamespace
 import configs.config as config
-from core.backtester import Backtester
+from backtester import Backtester
 
 def run_optimization():
     # ... (the main optimization loop is unchanged) ...
